+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Added untested Oculus Touch joystick bindings
 - Reduce seated position reset hold time from 3s to 2s
 - Reload the ED control bindings when a bindings file is edited
+- Read StartPreset.start and load default ED bindings if selected (built-in bindings only load after game startup)
 
 ## [0.7.0] - 2020-01-26